pyproject.toml
[project]
name = "androbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "appium-python-client>=5.1.1",
    "httpx>=0.28.1",
    "pillow>=11.2.1",
    "typer>=0.16.0",
    "requests",
    "fastmcp",
    "google-genai",
    "google-adk",
    "litellm",
    "gradio",
    "google-play-scraper",
    "supervision"
]
